# components/backend/src/main/resources/restaurant-api-doc.yaml
openapi: 3.0.3
info:
  title: Swagger Restaurant
  description: |-
    This is a sample Restaruant
  version: 1.0.1
...
servers:
  - url: http://localhost:8084
...
paths:
...
  /employee:
    get:
      tags:
        - employee
      summary: Get all employees
      operationId: getEmployees
      security:
      - sessionId: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'          
        '400':
          description: Invalid status value
    post:
      tags:
        - employee
      summary: Register an employee
      operationId: registerEmployee
      security:
      - sessionId: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found
        '405':
          description: Validation exception
  /customer:
    get:
      tags: 
        - customer
      summary: Get all customers
      operationId: listCustomers
      security:
      - sessionId: []
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: 
        - customer
      summary: Customer registration
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User successfully registered
        '400':
          description: Invalid request
  /customer/{customer-id}/booking:
    parameters:
      - name: customer-id
        in: path
        schema: { type: string }
        required: true
    get:
      tags:
        - customer
      summary: Query bookings for customer
      description: 
      operationId: queryCustomerBookings
      parameters:
        - name: from
          in: header
          schema:
            type: string
            format: date-time
          required: false
        - name: to
          in: header
          schema:
            type: string
            format: date-time
          required: false
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
                      
        '405':
          description: Invalid input
  /invoice/{orderId}:
    parameters:
      - name: orderId
        in: path
        description: ID of the order for which to generate an invoice
        required: true
        schema:
          type: string
          example: ObjectId('6544cd596955fe0a1c04fba9')
    get:
      tags:
        - customer
      summary: Get an invoice in PDF format by orderId
      operationId: getInvoice
      security:
      - sessionId: []
      responses:
        '200':
          description: Invoice PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Order not found
...
components:
  schemas:
    Table:
      required:
        - number
        - capacity
      type: object
      properties:
        number:
          type: integer
          format: int32
          example: 10
        capacity:
          type: integer
          format: int32
          example: 10
        status:
          type: string
          enum: [booked, free]
    Booking:
      required:
        - id
        - table
        - customer
        - startingDate
        - endingDate
        - status
        - peopleCount
      type: object
      properties:
        id:
          type: string
          example: ObjectId('6544cd596955fe0a1c04fba9')
        table:
          $ref: '#/components/schemas/Table'
        customer:
          $ref: '#/components/schemas/User'
        startingDate:
          type: string
          format: date-time
        endingDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, accepted]
        peopleCount:
          type: integer
...
    UserRegister:
      required:
        - name
        - email
        - mobil
        - address
        - password
      type: object
      properties:
        name:
          type: string
          example: Gipsz Jakap
        email:
          type: string
          example: gipsz@jakab.com
        mobil:
          type: string
          example: +36709834234
        address:
          type: string
          example: Budapest Lakatos utca 6.
        password:
          type: string
          example: password
...
    User:
      required:
        - name
        - email
        - mobil
        - address
      type: object
      properties:
        id:
          type: string
          example: ObjectId('6544cd596955fe0a1c04fba9')
        name:
          type: string
          example: Gipsz Jakap
        email:
          type: string
          example: gipsz@jakab.com
        mobil:
          type: string
          example: +36709834234
        address:
          type: string
          example: Budapest Lakatos utca 6.
...
  securitySchemes:
    sessionId:
      type: apiKey
      in: header
      name: sessionId